requirements.txt
bleach==1.5.0
dj-database-url==0.4.2
dj-static==0.0.6
Django==2.0.2
django-appconf==1.0.2
django-bootstrap4==0.0.6
django-classy-tags==0.8.0
django-geojson==2.11.0
django-imagekit==4.0.2
django-jquery==3.1.0
django-leaflet==0.23.0
django-mptt==0.8.7
django-nyt==1.0
django-sekizai==0.10.0
django-taggit==0.22.2
django-toolbelt==0.0.1
geopy==1.11.0
gunicorn==19.7.1
html5lib==0.9999999
Markdown==2.6.11
pilkit==2.0
Pillow==5.0.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
pytz==2018.3
six==1.11.0
sorl-thumbnail==12.4.1
static3==0.7.0
uWSGI==2.0.16
